refactor: complete refactor of solved questions done
