--- README.md
@@ -1,5 +1,0 @@
-## binarysearch 
-
-Solving some binary Search questions for fun and profit.
-
-Read my [blog](https://github.com/rohan3kishibe/blog/) on how I am solving these questions.
