+++ app/src/main/java/binsearch_java/App.java
@@ -1,0 +1,11 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package binsearch_java;
+
+public class App {
+
+    public static void main(String[] args) {
+        System.out.println("Currently not running");
+    }
+}
