+++ app/src/main/java/binsearch_java/tests.md
@@ -1,0 +1,20 @@
+# Tests for Code that have already passed:
+1.  public static void runq151() {
+        System.out.println(new q151().solve(9, 5));
+    }
+2.  public static void runq5() {
+        System.out.println(new q5().solve(121));
+        System.out.println(new q5().solve(20200202));
+        System.out.println(new q5().solve(43));
+    }
+3.  public static void runq750() {
+        System.out.println(new q750().solve(25));
+    }
+4.  public static void runq85() {
+        System.out.println(new q85().solve(123));
+        System.out.println(new q85().solve(50));
+    }
+5.  public static void runq982() {
+        System.out.println(new q982().solve([15,10,30,50,25],20));
+        System.out.println(new q85().solve([15,10,30,50,25],51));
+    }
